Fixes around ExternalOutput refactor
* Function syncClose returns if not recording.
* Properly check if context_ isn't null before free it.
* Do not exit if context couldn't be initialized.
* Remove not used save jpeg function from Coder class.
* Add CoderCodec isInitialized getter.
* Add CoderCodec width/height accessors.
* Add CoderCodec abitlity to initialize AVStream.
* Rename InitContextCB to InitContextBeforeOpenCB.
* Fix Coder::initContext return variable was overrided.
* InitEncoder / InitDecord now returns bool.
* Protect AVCodec in CoderCodec.
* Check ffmpeg is installed before download and compile.

+++ scripts/installUbuntuDeps.sh
@@ -157,12 +157,16 @@ install_mediadeps(){
   sudo apt-get -qq install yasm libvpx. libx264.
   if [ -d $LIB_DIR ]; then
     cd $LIB_DIR
-    curl -O -L https://ffmpeg.org/releases/ffmpeg-3.4.2.tar.gz
-    tar -zxvf ffmpeg-3.4.2.tar.gz
-    cd ffmpeg-3.4.2
-    PKG_CONFIG_PATH=${PREFIX_DIR}/lib/pkgconfig ./configure --prefix=$PREFIX_DIR --enable-shared --enable-gpl --enable-libvpx --enable-libx264 --enable-libopus --disable-doc --enable-avresample
-    make $FAST_MAKE -s V=0
-    make install
+    if [ ! -f ./ffmpeg-3.4.2.tar.gz ]; then
+      curl -O -L https://ffmpeg.org/releases/ffmpeg-3.4.2.tar.gz
+      tar -zxvf ffmpeg-3.4.2.tar.gz
+      cd ffmpeg-3.4.2
+      PKG_CONFIG_PATH=${PREFIX_DIR}/lib/pkgconfig ./configure --prefix=$PREFIX_DIR --enable-shared --enable-gpl --enable-libvpx --enable-libx264 --enable-libopus --disable-doc --enable-avresample
+      make $FAST_MAKE -s V=0
+      make install
+    else
+      echo "ffmpeg already installed"
+    fi
     cd $CURRENT_DIR
   else
     mkdir -p $LIB_DIR
@@ -176,12 +180,16 @@ install_mediadeps_nogpl(){
   sudo apt-get -qq install yasm libvpx.
   if [ -d $LIB_DIR ]; then
     cd $LIB_DIR
-    curl -O -L https://ffmpeg.org/releases/ffmpeg-3.4.2.tar.gz
-    tar -zxvf ffmpeg-3.4.2.tar.gz
-    cd ffmpeg-3.4.2
-    PKG_CONFIG_PATH=${PREFIX_DIR}/lib/pkgconfig ./configure --prefix=$PREFIX_DIR --enable-shared --enable-gpl --enable-libvpx --enable-libopus --disable-doc --enable-avresample
-    make $FAST_MAKE -s V=0
-    make install
+    if [ ! -f ./ffmpeg-3.4.2.tar.gz ]; then
+      curl -O -L https://ffmpeg.org/releases/ffmpeg-3.4.2.tar.gz
+      tar -zxvf ffmpeg-3.4.2.tar.gz
+      cd ffmpeg-3.4.2
+      PKG_CONFIG_PATH=${PREFIX_DIR}/lib/pkgconfig ./configure --prefix=$PREFIX_DIR --enable-shared --enable-gpl --enable-libvpx --enable-libopus --disable-doc --enable-avresample
+      make $FAST_MAKE -s V=0
+      make install
+    else
+      echo "ffmpeg already installed"
+    fi
     cd $CURRENT_DIR
   else
     mkdir -p $LIB_DIR
